Move the Output code into a io package
